Minor Fixes and changes

+++ app/src/main/java/com/example/proyecto/models/HServicesScheduleData.kt
@@ -5,7 +5,7 @@ data class HServicesScheduleData(
     val created_by_name: String,
     val day_name: String,
     val day_of_week: Int,
-    val duration_hours: Int,
+    val duration_hours: Float,
     val end_time: String,
     val id: String,
     val is_available: Boolean,
